requirements.txt
flask
boto3 
pyyaml